Add runtime integrity self-testing of wordlists
The code now self-tests the SHA-256 hashes of the wordlists with which
it was compiled.  I am a bit nervous about build-time bugs in platform
shell tools which might behave differently in different locales, etc.
Will *your* platform's shell tools process UTF-8 strings without
mangling?  Now, you can be sure with `make check`.

This test currently runs by default.  It is slow, so I will probably
disable it when the undocumented -T flag is not used.  I do think the
proportion of test code to feature code is growing appropriately, for a
utility which pertains to Other People's Money.

I also added code to reproduce a wordlist on stdout, in a format
identical to the source file.  However, that function cannot yet be used
due to lack of language selection UI.

The language selection UI is next!