Unlist proposed Indonesian from -L option
Buglet:  A wordlist not yet accepted into the BIP repository should not
be visible in the user interface.  It is included only for purposes of
testing, since it may be changed or deleted at any time.  The -L option
automatically listed *all* available wordlist languages; now it has a
special-case check to not list Indonesian.  Change this back if/when
BIP PR https://github.com/bitcoin/bips/pull/621 be accepted.