Overhaul test system for multilingual vectors
The essential code is now in place to self-test with vectors in multiple
languages; and Japanese test vectors have been added to the repository.
However, the code is not yet enabled.

The Japanese test vector designer properly included mnemonic phrases
which are not normalized to Unicode NFKD.  Therefore, with the current
code, Japanese test vectors with these indices fail:  0-7, 9, 10, 12-23.
Tests 8 and 11 succeed.

The multi-language test vectors will be re-enabled (and enhanced) after
Unicode normalization support is added, which will soon be done anyway
to prepare for seed generation.

The English test vectors were also added from the Japanese repository;
whereupon I discovered that this set and the set from the Trezor
repository overlap, but are not identical.  The two files, vectors.json
and test_EN_BIP39.json with 24 vectors each, are now combined and
deduplicated into a single set of 36 vectors.  Deduplication is keyed
exclusively on the input entropy bitstring; I did not affirmatively
verify that both sets contain identical results for identical inputs.