Add "baked" version of vectors.c
The generation of vectors.c from JSON sources currently uses jq (from
within what may be the worst shell script I have ever written).  jq is
not currently installed on my airgap machine; and it may not be on
yours, either.  I demand that tool dependencies for an ordinary build be
restricted to a standard C/Unix toolchain.  Adding to or modifying the
test vectors will, however, require jq.