Add proposed Czech wordlist via bitcoin/bips#493
Notice:  This is hidden behind the -W flag; see 8aaa6f3.

This is not exactly the wordlist proposed in the pull request.  It is
the czech.txt from zizelevak/bips@b7f682f, as modified by approximately
the following command:

	uconv -f utf-8 -t utf-8 -x '::nfkd;' < czech.txt | \
		LC_ALL=C LANG=C sort -s > normalized/czech.txt

The *result* has been confirmed to not have any leading BOM, and to have
a final line terminated with '\n' (bitcoin/bips#622).  I did not yet
examine the source for these issues.  I did examine the source to
confirm that no lines had any trailing whitespace (see 08a05b4).

SHA-256 hash for the *resulting* czech.txt:
195136b3ba0f3099a9df625e0963f4efb56625b91c3a76bc5b4a9466a26880f7