Add passphrase entry backend
easyseed now features backend code for reading a passphrase from the
terminal via the BSD readpassphrase(3) API.  On Linux, this introduces a
dependency on libbsd for those who need to type in a passphrase.

Passphrase validation code has also been added.  Input validation, first
and foremost!  The validation code is designed to sanity-check
passphrases read from the tty, or (in the future) frmo a file, fifo,
etc.  Also in the future, the same validation path can be used for
mnemonics entered by the user for the purpose of generating keys for
wallet restoration.

The -p option has been perfunctorily added; but as yet, it does nothing.
This functionality is not yet exposed to the user.