Add undocumented UI for first full functionality
easyseed is now essentially feature-complete for BIP 39, and has some
functionality from BIP 32.  Generation of output seed (with or without
passphrase) and BIP 32 "xprv" extended master private key are available
through the following options.  The manpage is not yet updated, pending
further testing and investigation of the issue described below under -E.

-A	All.  Do all and everything.  Output is on stdout in a format
human-readable, but not difficult for simple scripts to parse.  Provided
information:  The input entropy in hex, the mnemonic, the BIP 39 seed in
hex, and the "xprv" BIP 32 extended master private key.

-p	Use a passphrase, entered and confirmed from the terminal.
Requires a BSD system, or libbsd on Linux.

(-E	Electrum output.  Use nonstandard "xprv" version numbers to also
output "yprv" and "zprv".  See the Electrum 3.0 release notes for an
explanation, and electrum/lib/bitcoin.py for the magic numbers.  I added
this for the purpose of making backward-compatible Segwit P2WPKH-in-P2SH
easy to use in Electrum, the most popular desktop light client.
However, this is *broken* insofar as the wallets created with "xprv",
"yprv", and "zprv" respectively generate different addresses than those
restored in Electrum with the BIP 39 mnemonic and a correct derivation
path.  This is tested with "xprv" versus mnemonic with path m/44'/0'/0',
"yprv" versus mnemnic with path m/49'/0'/0', and "zprv" versus mnemonic
with path m/84'/0'/0' (although I did not expect this last to work,
because BIP 84 is a draft first published 2017-12-28).  The mnemonic may
be used, *or* the "xprv"/"yprv" may be used; but attempts to the
mnemonic as a backup for a wallet created with "xprv"/"yprv"/"zprv" may
result in user loss of funds.)

I do not see how this described problem could be a bug in my code, which
passes all test vectors.  I manually tried a few of the test vectors
with Electrum, using the "xprv" to generate one wallet, and the mnemonic
plus passphrase ("extra words") plus derivation path m/44'/0'/0' to
generate another wallet.  The wallets contained different addresses.

The -E option may or may not be removed.  The -A and -p options will
stay; and others will be added for finer control.

Passphrase entry on Linux with libbsd was broken as added in 37defc2.
Now that there is means to test it, I noticed and fixed the bug.