Release full BIP 39 functionality; update manpage
In addition to previously undocumented features from c912169 and
37defc2, this commit does:

- Update the manpage to document new functionality, for full BIP 39
support and rudimentary BIP 32 support.

- Add option -j to read passphrase from file.  This is principally
intended for use with FIFOs in scripts, for generation of test vectors.

- Modestly refactor a few parts needed to make all passphrase
functionality consistent, and tighten up handling of options in main().

- Refuse to read -k keymat or -j passphrases from a tty.

- Add the undocumented -D option, which at present only causes the -A
output to include the passphrase (if any).  This is for generation of
test vectors.

- Hide the undocumented -E option behind an undocumented compile-time
knob, to prevent potential user funds loss.  See spesmilo/electrum#3671.

- Add a caveat to the manpage against using the output xprv with
Electrum.  See again, spesmilo/electrum#3671.

- Update the README.md to advertise new functionality.